refactor(core): split index/mod.rs, graph/algorithms/traversal.rs, graph/clustering.rs into submodules
Pure code-move refactor (wave 2, raw-line metric) — zero logic changes,
facade mod.rs re-exports preserve all paths.

- index/mod.rs (1797 -> 179) -> +dist,label_index,knn_index,
  property_index — 49/49 tests preserved
- graph/algorithms/traversal.rs (1715) -> traversal/{mod,bfs_dfs,
  shortest_path,components,centrality,similarity,mst}
- graph/clustering.rs (1669) -> clustering/ — 24/24 tests preserved

Part of phase5_split-oversized-files-wave2 (rulebook task).
