+++ .rulebook/tasks/README.md
@@ -2,15 +2,15 @@
 
 > Auto-generated by rulebook. Do not edit manually.
 
-**Total**: 14 tasks | **Completed**: 0 | **In Progress**: 1 | **Pending**: 13
+**Total**: 14 tasks | **Completed**: 1 | **In Progress**: 1 | **Pending**: 12
 
 ## phase5
 
 | Status | Task | Progress | Description |
 |--------|------|----------|-------------|
 | ⬚ | phase5_hub-integration | 10/87 | hub integration |
-| 🔄 | phase5_split-oversized-files | 19/23 | split oversized files |
-| ⬚ | phase5_split-oversized-files-wave2 | 0/5 | split oversized files wave2 |
+| ✅ | phase5_split-oversized-files | 23/23 | split oversized files |
+| 🔄 | phase5_split-oversized-files-wave2 | 0/13 | split oversized files wave2 |
 | ⬚ | phase5_wire-or-remove-dead-integration-test | 0/7 | wire or remove dead integration test |
 
 ## phase6
