+++ .rulebook/tasks/phase5_split-oversized-files-wave2/.metadata.json
@@ -1,5 +1,5 @@
 {
-  "status": "pending",
+  "status": "in-progress",
   "createdAt": "2026-06-09T23:34:01.744Z",
-  "updatedAt": "2026-06-09T23:34:01.744Z"
+  "updatedAt": "2026-06-10T09:10:45.705Z"
 }
