+++ .rulebook/tasks/phase5_split-oversized-files/.metadata.json
@@ -1,5 +1,5 @@
 {
-  "status": "in-progress",
+  "status": "completed",
   "createdAt": "2026-06-09T21:12:28.487Z",
-  "updatedAt": "2026-06-09T21:13:43.761Z"
+  "updatedAt": "2026-06-09T23:41:31.124Z"
 }
