+++ crates/nexus-core/src/graph/clustering/grouping.rs
@@ -1,0 +1,127 @@
+//! Grouping-style clustering: hierarchical, label-based, and property-based.
+
+use crate::error::Result;
+use crate::graph::simple::{Graph, Node, PropertyValue};
+
+use super::engine::ClusteringEngine;
+use super::types::{Cluster, ClusteringResult, LinkageType};
+
+impl ClusteringEngine {
+    /// Hierarchical clustering implementation
+    pub(super) fn hierarchical_clustering(
+        &self,
+        graph: &Graph,
+        _linkage: &LinkageType,
+        nodes: &[Node],
+    ) -> Result<ClusteringResult> {
+        // Simplified hierarchical clustering implementation
+        let features = self.extract_features(graph, nodes)?;
+        let n = features.len();
+
+        // Start with each node as its own cluster
+        let clusters: Vec<Cluster> = (0..n)
+            .map(|i| Cluster::new(i as u64, vec![nodes[i].id]))
+            .collect();
+
+        // For now, just return the initial clusters
+        // A full implementation would perform the hierarchical merging
+        let metrics = self.calculate_metrics(graph, &clusters, &features, nodes)?;
+
+        Ok(ClusteringResult {
+            clusters,
+            algorithm: self.config.algorithm.clone(),
+            iterations: 0,
+            converged: true,
+            metrics,
+        })
+    }
+
+    /// Label-based grouping
+    pub(super) fn label_based_grouping(
+        &self,
+        graph: &Graph,
+        nodes: &[Node],
+    ) -> Result<ClusteringResult> {
+        use crate::graph::simple::NodeId;
+        use std::collections::HashMap;
+
+        let mut label_groups: HashMap<String, Vec<NodeId>> = HashMap::new();
+
+        for node in nodes {
+            for label in &node.labels {
+                label_groups.entry(label.clone()).or_default().push(node.id);
+            }
+        }
+
+        let clusters: Vec<Cluster> = label_groups
+            .into_iter()
+            .enumerate()
+            .map(|(i, (label, node_ids))| {
+                let mut cluster = Cluster::new(i as u64, node_ids);
+                cluster.set_metadata("label".to_string(), PropertyValue::String(label));
+                cluster
+            })
+            .collect();
+
+        let features = self.extract_features(graph, nodes)?;
+        let metrics = self.calculate_metrics(graph, &clusters, &features, nodes)?;
+
+        Ok(ClusteringResult {
+            clusters,
+            algorithm: self.config.algorithm.clone(),
+            iterations: 0,
+            converged: true,
+            metrics,
+        })
+    }
+
+    /// Property-based grouping
+    pub(super) fn property_based_grouping(
+        &self,
+        graph: &Graph,
+        property_key: &str,
+        nodes: &[Node],
+    ) -> Result<ClusteringResult> {
+        use crate::graph::simple::NodeId;
+        use std::collections::HashMap;
+
+        let mut property_groups: HashMap<String, Vec<NodeId>> = HashMap::new();
+
+        for node in nodes {
+            let group_key = if let Some(value) = node.get_property(property_key) {
+                match value {
+                    PropertyValue::String(s) => s.clone(),
+                    PropertyValue::Int64(i) => i.to_string(),
+                    PropertyValue::Float64(f) => f.to_string(),
+                    PropertyValue::Bool(b) => b.to_string(),
+                    _ => "unknown".to_string(),
+                }
+            } else {
+                "null".to_string()
+            };
+
+            property_groups.entry(group_key).or_default().push(node.id);
+        }
+
+        let clusters: Vec<Cluster> = property_groups
+            .into_iter()
+            .enumerate()
+            .map(|(i, (value, node_ids))| {
+                let mut cluster = Cluster::new(i as u64, node_ids);
+                cluster.set_metadata(property_key.to_string(), PropertyValue::String(value));
+                cluster
+            })
+            .collect();
+
+        let features = self.extract_features(graph, nodes)?;
+        let metrics = self.calculate_metrics(graph, &clusters, &features, nodes)?;
+
+        Ok(ClusteringResult {
+            clusters,
+            algorithm: self.config.algorithm.clone(),
+            iterations: 0,
+            converged: true,
+            metrics,
+        })
+    }
+}
