+++ crates/nexus-core/src/index/mod.rs
@@ -6,63 +6,26 @@
 //! - Full-text index: Tantivy per label/key
 //! - KNN index: Simple cosine similarity for MVP
 
-use crate::simd;
-use crate::{Error, Result};
-use hnsw_rs::prelude::*;
-use parking_lot::RwLock;
-use roaring::RoaringBitmap;
-use std::collections::{BTreeMap, HashMap, HashSet};
-use std::sync::Arc;
+use crate::Result;
 
 pub mod btree;
 pub mod composite_btree;
+pub mod dist;
 pub mod fulltext;
 pub mod fulltext_analyzer;
 pub mod fulltext_registry;
 pub mod fulltext_writer;
+pub mod knn_index;
+pub mod label_index;
 pub mod pending_updates;
+pub mod property_index;
 pub mod rtree;
 
-/// Default dimensionality for `KnnIndex` when no per-call override is
-/// supplied.
-///
-/// 128 matches the output dimension of the smallest of the HiveLLM
-/// vectorizer profiles (see `hivellm/vectorizer`). It is also the
-/// default MiniLM dim, so SDK examples that embed with a generic
-/// off-the-shelf model round-trip without an explicit dimension
-/// argument. Tests that exercise specific dimensions (64, 256, 768,
-/// 1536) still pass their dimension explicitly to
-/// [`KnnIndex::new_default`] / [`KnnIndex::new`]; this constant only
-/// applies to the "give me a default index" path.
-pub const DEFAULT_VECTORIZER_DIMENSION: usize = 128;
-
-/// SIMD-backed cosine distance plugged into HNSW.
-///
-/// Delegates to `crate::simd::distance::cosine_f32`, which resolves
-/// to the fastest kernel the host CPU supports (AVX-512 → AVX2 → SSE4.2
-/// → NEON → Scalar). Produces bit-equivalent results to `DistSimdCosine`
-/// within the `1e-4` tolerance documented in ADR-003.
-#[derive(Default, Copy, Clone)]
-pub struct DistSimdCosine;
-
-impl Distance<f32> for DistSimdCosine {
-    fn eval(&self, va: &[f32], vb: &[f32]) -> f32 {
-        simd::distance::cosine_f32(va, vb)
-    }
-}
-
-/// SIMD-backed squared L2 distance (Euclidean²) plugged into HNSW.
-#[derive(Default, Copy, Clone)]
-pub struct DistSimdL2;
-
-impl Distance<f32> for DistSimdL2 {
-    fn eval(&self, va: &[f32], vb: &[f32]) -> f32 {
-        simd::distance::l2_sq_f32(va, vb)
-    }
-}
-
-/// Type alias for property index trees
-type PropertyIndexTree = BTreeMap<PropertyValue, RoaringBitmap>;
+// Re-export everything that was previously reachable at `crate::index::*`
+pub use dist::{DEFAULT_VECTORIZER_DIMENSION, DistSimdCosine, DistSimdL2};
+pub use knn_index::{KnnConfig, KnnIndex, KnnIndexStats};
+pub use label_index::{LabelIndex, LabelIndexStats};
+pub use property_index::{PropertyIndex, PropertyIndexStats, PropertyValue};
 
 /// Index manager that coordinates all index types
 #[derive(Clone)]
@@ -152,1110 +115,9 @@ impl IndexManager {
     }
 }
 
-/// Label bitmap index using roaring bitmaps
-///
-/// Maps label_id → bitmap of node_ids for fast label-based queries.
-/// Uses RoaringBitmap for efficient compression and operations.
-#[derive(Clone)]
-pub struct LabelIndex {
-    /// Mapping from label_id to bitmap of node_ids
-    label_bitmaps: Arc<RwLock<HashMap<u32, RoaringBitmap>>>,
-}
-
-/// Statistics for label index
-#[derive(Debug, Clone, Default)]
-pub struct LabelIndexStats {
-    /// Total number of nodes indexed
-    pub total_nodes: u64,
-    /// Number of unique labels
-    pub label_count: u32,
-    /// Average nodes per label
-    pub avg_nodes_per_label: f64,
-}
-
-impl LabelIndex {
-    fn recompute_stats(bitmaps: &HashMap<u32, RoaringBitmap>) -> LabelIndexStats {
-        let mut unique_nodes: HashSet<u32> = HashSet::new();
-        let mut total_entries: u64 = 0;
-
-        for bitmap in bitmaps.values() {
-            total_entries += bitmap.len();
-            unique_nodes.extend(bitmap.iter());
-        }
-
-        let label_count = bitmaps.len() as u32;
-        let avg_nodes_per_label = if label_count > 0 {
-            total_entries as f64 / label_count as f64
-        } else {
-            0.0
-        };
-
-        LabelIndexStats {
-            total_nodes: unique_nodes.len() as u64,
-            label_count,
-            avg_nodes_per_label,
-        }
-    }
-
-    /// Create a new label index
-    pub fn new() -> Self {
-        Self {
-            label_bitmaps: Arc::new(RwLock::new(HashMap::new())),
-        }
-    }
-
-    /// Add a node with given labels
-    pub fn add_node(&self, node_id: u64, label_ids: &[u32]) -> Result<()> {
-        let mut bitmaps = self.label_bitmaps.write();
-
-        for &label_id in label_ids {
-            bitmaps.entry(label_id).or_default().insert(node_id as u32);
-        }
-
-        Ok(())
-    }
-
-    /// Remove a node from all labels
-    pub fn remove_node(&self, node_id: u64) -> Result<()> {
-        let mut bitmaps = self.label_bitmaps.write();
-
-        for bitmap in bitmaps.values_mut() {
-            bitmap.remove(node_id as u32);
-        }
-
-        bitmaps.retain(|_, bitmap| !bitmap.is_empty());
-        Ok(())
-    }
-
-    /// Replace the labels associated with a node
-    pub fn set_node_labels(&self, node_id: u64, label_ids: &[u32]) -> Result<()> {
-        let mut bitmaps = self.label_bitmaps.write();
-
-        for bitmap in bitmaps.values_mut() {
-            bitmap.remove(node_id as u32);
-        }
-
-        for &label_id in label_ids {
-            bitmaps.entry(label_id).or_default().insert(node_id as u32);
-        }
-
-        bitmaps.retain(|_, bitmap| !bitmap.is_empty());
-        Ok(())
-    }
-
-    /// Get all nodes with a specific label
-    pub fn get_nodes(&self, label_id: u32) -> Result<RoaringBitmap> {
-        let bitmaps = self.label_bitmaps.read();
-        Ok(bitmaps.get(&label_id).cloned().unwrap_or_default())
-    }
-
-    /// Get nodes that have ALL specified labels (intersection)
-    pub fn get_nodes_with_labels(&self, label_ids: &[u32]) -> Result<RoaringBitmap> {
-        if label_ids.is_empty() {
-            return Ok(RoaringBitmap::new());
-        }
-
-        let bitmaps = self.label_bitmaps.read();
-        let mut result = bitmaps.get(&label_ids[0]).cloned().unwrap_or_default();
-
-        for &label_id in &label_ids[1..] {
-            if let Some(bitmap) = bitmaps.get(&label_id) {
-                result &= bitmap;
-            } else {
-                return Ok(RoaringBitmap::new());
-            }
-        }
-
-        Ok(result)
-    }
-
-    /// Get nodes that have ANY of the specified labels (union)
-    pub fn get_nodes_with_any_labels(&self, label_ids: &[u32]) -> Result<RoaringBitmap> {
-        let bitmaps = self.label_bitmaps.read();
-        let mut result = RoaringBitmap::new();
-
-        for &label_id in label_ids {
-            if let Some(bitmap) = bitmaps.get(&label_id) {
-                result |= bitmap;
-            }
-        }
-
-        Ok(result)
-    }
-
-    /// Estimate cardinality for a label
-    pub fn estimate_cardinality(&self, label_id: u32) -> u64 {
-        let bitmaps = self.label_bitmaps.read();
-        bitmaps.get(&label_id).map(|b| b.len()).unwrap_or(0)
-    }
-
-    /// Get statistics. Computed on demand from the label bitmaps — the
-    /// previous design recomputed (O(N) over every node in every bitmap) on
-    /// every `add_node`/`remove_node`, which pinned CPU under sustained write
-    /// load (issue #12). Stats are diagnostic-only, so the cost moves to the
-    /// (rare) read instead of the (hot) write.
-    pub fn get_stats(&self) -> LabelIndexStats {
-        let bitmaps = self.label_bitmaps.read();
-        Self::recompute_stats(&bitmaps)
-    }
-
-    /// Check if a label exists
-    pub fn has_label(&self, label_id: u32) -> bool {
-        let bitmaps = self.label_bitmaps.read();
-        bitmaps.contains_key(&label_id)
-    }
-
-    /// Get all label IDs
-    pub fn get_all_labels(&self) -> Vec<u32> {
-        let bitmaps = self.label_bitmaps.read();
-        bitmaps.keys().copied().collect()
-    }
-
-    /// Clear all data
-    pub fn clear(&mut self) -> Result<()> {
-        let mut bitmaps = self.label_bitmaps.write();
-        bitmaps.clear();
-
-        Ok(())
-    }
-
-    /// Health check for the label index
-    pub fn health_check(&self) -> Result<()> {
-        let bitmaps = self.label_bitmaps.read();
-
-        // Check if the number of labels is reasonable
-        if bitmaps.len() > 1_000_000 {
-            // 1 million max
-            return Err(Error::index("Too many labels"));
-        }
-
-        // Check if the total nodes count is reasonable
-        let total_nodes = Self::recompute_stats(&bitmaps).total_nodes;
-        if total_nodes > 1_000_000_000 {
-            // 1 billion max
-            return Err(Error::index("Too many nodes in label index"));
-        }
-
-        Ok(())
-    }
-}
-
-impl Default for LabelIndex {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-/// Configuration for an HNSW-backed KNN index.
-///
-/// HNSW keeps its graph and vector data resident in RAM, so `max_elements`
-/// directly bounds memory footprint. A single `KnnIndex` with 10_000 slots
-/// and 128-dim f32 vectors consumes ~15 MB; multiply by number of labels
-/// that get their own index.
-#[derive(Debug, Clone, Copy)]
-pub struct KnnConfig {
-    /// Upper bound on vectors stored in the index. HNSW allocates this
-    /// capacity eagerly, so keep it close to the expected working-set size.
-    pub max_elements: usize,
-    /// Maximum number of outgoing connections per node (HNSW `M`).
-    pub max_connections: usize,
-    /// Maximum layer count in the HNSW hierarchy.
-    pub max_layer: usize,
-    /// Size of the dynamic candidate list during graph construction.
-    pub ef_construction: usize,
-}
-
-impl Default for KnnConfig {
-    fn default() -> Self {
-        // Previous hardcoded values reserved 10_000 slots unconditionally,
-        // which is expensive for small deployments that only create a handful
-        // of vectors. Callers with larger working sets should construct a
-        // KnnConfig explicitly rather than relying on the default.
-        Self {
-            max_elements: 1_000,
-            max_connections: 16,
-            max_layer: 16,
-            ef_construction: 200,
-        }
-    }
-}
-
-/// KNN vector index using HNSW (Hierarchical Navigable Small World)
-///
-/// Maps node_id → embedding for fast similarity search.
-/// Uses HNSW algorithm for sub-linear search complexity.
-#[derive(Clone)]
-pub struct KnnIndex {
-    /// HNSW index for fast KNN search
-    hnsw: Arc<RwLock<Hnsw<'static, f32, DistSimdCosine>>>,
-    /// Mapping from node_id to vector index in HNSW
-    node_to_index: Arc<RwLock<HashMap<u64, usize>>>,
-    /// Mapping from vector index to node_id
-    index_to_node: Arc<RwLock<HashMap<usize, u64>>>,
-    /// Vector dimension
-    dimension: usize,
-    /// HNSW configuration — retained so `clear()` can recreate the index
-    /// with the same parameters the caller chose at construction time.
-    config: KnnConfig,
-    /// Statistics
-    stats: Arc<RwLock<KnnIndexStats>>,
-    /// Next available index
-    next_index: Arc<RwLock<usize>>,
-}
-
-/// Statistics for KNN index
-#[derive(Debug, Clone, Default)]
-pub struct KnnIndexStats {
-    /// Total number of vectors indexed
-    pub total_vectors: u64,
-    /// Vector dimension
-    pub dimension: usize,
-    /// Average search time in microseconds
-    pub avg_search_time_us: f64,
-}
-
-impl KnnIndex {
-    /// Create a new KNN index with a caller-supplied HNSW configuration.
-    ///
-    /// # Arguments
-    /// * `dimension` - Vector dimension (must be > 0 and <= 4096)
-    /// * `config`    - HNSW parameters; see [`KnnConfig`]
-    ///
-    /// # Errors
-    /// Returns an error if dimension is invalid
-    pub fn with_config(dimension: usize, config: KnnConfig) -> Result<Self> {
-        if dimension == 0 || dimension > 4096 {
-            return Err(Error::InvalidId(format!(
-                "Invalid vector dimension: {}",
-                dimension
-            )));
-        }
-
-        let hnsw = Hnsw::new(
-            config.max_connections,
-            config.max_elements,
-            config.max_layer,
-            config.ef_construction,
-            DistSimdCosine,
-        );
-
-        Ok(Self {
-            hnsw: Arc::new(RwLock::new(hnsw)),
-            node_to_index: Arc::new(RwLock::new(HashMap::new())),
-            index_to_node: Arc::new(RwLock::new(HashMap::new())),
-            dimension,
-            config,
-            stats: Arc::new(RwLock::new(KnnIndexStats {
-                total_vectors: 0,
-                dimension,
-                avg_search_time_us: 0.0,
-            })),
-            next_index: Arc::new(RwLock::new(0)),
-        })
-    }
-
-    /// Create a new KNN index with default HNSW parameters.
-    ///
-    /// Equivalent to `KnnIndex::with_config(dimension, KnnConfig::default())`.
-    /// Callers that need more capacity should construct [`KnnConfig`] explicitly.
-    pub fn new(dimension: usize) -> Result<Self> {
-        Self::with_config(dimension, KnnConfig::default())
-    }
-
-    /// Create a new KNN index with default parameters (alias of [`KnnIndex::new`]).
-    pub fn new_default(dimension: usize) -> Result<Self> {
-        Self::new(dimension)
-    }
-
-    /// Add a vector for a node
-    pub fn add_vector(&self, node_id: u64, embedding: Vec<f32>) -> Result<()> {
-        if embedding.len() != self.dimension {
-            return Err(Error::InvalidId(format!(
-                "Vector dimension mismatch: expected {}, got {}",
-                self.dimension,
-                embedding.len()
-            )));
-        }
-
-        let hnsw = self.hnsw.write();
-        let mut node_to_index = self.node_to_index.write();
-        let mut index_to_node = self.index_to_node.write();
-        let mut next_index = self.next_index.write();
-
-        // Check if node already exists
-        if let Some(&_existing_index) = node_to_index.get(&node_id) {
-            // Update existing vector - HNSW doesn't support updates, so we'll just add new
-            // In a production system, you might want to implement a more sophisticated update mechanism
-        }
-
-        // Add new vector to HNSW using insert method
-        let vector_index = *next_index;
-        hnsw.insert((&embedding, vector_index));
-
-        // Update mappings
-        node_to_index.insert(node_id, vector_index);
-        index_to_node.insert(vector_index, node_id);
-        *next_index += 1;
-
-        // Update statistics
-        let mut stats = self.stats.write();
-        stats.total_vectors += 1;
-
-        Ok(())
-    }
-
-    /// Remove a vector for a node
-    pub fn remove_vector(&self, node_id: u64) -> Result<()> {
-        let mut node_to_index = self.node_to_index.write();
-        let mut index_to_node = self.index_to_node.write();
-
-        if let Some(&vector_index) = node_to_index.get(&node_id) {
-            // Remove from mappings
-            node_to_index.remove(&node_id);
-            index_to_node.remove(&vector_index);
-
-            // Update statistics
-            let mut stats = self.stats.write();
-            stats.total_vectors = stats.total_vectors.saturating_sub(1);
-        }
-
-        Ok(())
-    }
-
-    /// Default `ef` (size of the dynamic candidate list at search time).
-    /// Larger values trade latency for recall.
-    pub const DEFAULT_EF_SEARCH: usize = 50;
-
-    /// Search for k nearest neighbors using cosine similarity.
-    ///
-    /// Uses [`KnnIndex::DEFAULT_EF_SEARCH`] as the HNSW `ef` parameter.
-    /// For tunable recall/latency tradeoffs, use
-    /// [`KnnIndex::search_knn_with_ef`].
-    pub fn search_knn(&self, query: &[f32], k: usize) -> Result<Vec<(u64, f32)>> {
-        self.search_knn_with_ef(query, k, Self::DEFAULT_EF_SEARCH)
-    }
-
-    /// Search for k nearest neighbors with an explicit HNSW `ef` parameter.
-    ///
-    /// `ef_search` controls the size of the dynamic candidate list during
-    /// the descent. The HNSW algorithm requires `ef_search >= k`; values
-    /// below `k` are silently raised to `k` so the caller never gets
-    /// fewer results than requested.
-    ///
-    /// Recall and latency both grow with `ef_search`. Typical sweep
-    /// ranges for production tuning are `ef_search ∈ {50, 100, 200, 400}`
-    /// — see `crates/nexus-knn-bench` for the full Pareto methodology.
-    pub fn search_knn_with_ef(
-        &self,
-        query: &[f32],
-        k: usize,
-        ef_search: usize,
-    ) -> Result<Vec<(u64, f32)>> {
-        if query.len() != self.dimension {
-            return Err(Error::InvalidId(format!(
-                "Query dimension mismatch: expected {}, got {}",
-                self.dimension,
-                query.len()
-            )));
-        }
-
-        let ef = ef_search.max(k);
-        let start_time = std::time::Instant::now();
-
-        let hnsw = self.hnsw.read();
-        let index_to_node = self.index_to_node.read();
-
-        let search_results = hnsw.search(query, k, ef);
-
-        let mut results = Vec::new();
-        for neighbour in search_results {
-            if let Some(&node_id) = index_to_node.get(&neighbour.d_id) {
-                let similarity = 1.0 - neighbour.distance;
-                results.push((node_id, similarity));
-            }
-        }
-
-        let search_time_us = start_time.elapsed().as_micros() as f64;
-        let mut stats = self.stats.write();
-        stats.avg_search_time_us = (stats.avg_search_time_us + search_time_us) / 2.0;
-
-        Ok(results)
-    }
-
-    /// Search for k nearest neighbors with default k=10
-    pub fn search_knn_default(&self, query: &[f32]) -> Result<Vec<(u64, f32)>> {
-        self.search_knn(query, 10)
-    }
-
-    /// Get statistics
-    pub fn get_stats(&self) -> KnnIndexStats {
-        self.stats.read().clone()
-    }
-
-    /// Get vector dimension
-    pub fn dimension(&self) -> usize {
-        self.dimension
-    }
-
-    /// Check if a node has a vector
-    pub fn has_vector(&self, node_id: u64) -> bool {
-        let node_to_index = self.node_to_index.read();
-        node_to_index.contains_key(&node_id)
-    }
-
-    /// Get all node IDs with vectors
-    pub fn get_all_nodes(&self) -> Vec<u64> {
-        let node_to_index = self.node_to_index.read();
-        node_to_index.keys().copied().collect()
-    }
-
-    /// Clear all data
-    pub fn clear(&mut self) -> Result<()> {
-        let mut hnsw = self.hnsw.write();
-        let mut node_to_index = self.node_to_index.write();
-        let mut index_to_node = self.index_to_node.write();
-        let mut next_index = self.next_index.write();
-
-        // Recreate the HNSW index using the config this instance was built with.
-        *hnsw = Hnsw::new(
-            self.config.max_connections,
-            self.config.max_elements,
-            self.config.max_layer,
-            self.config.ef_construction,
-            DistSimdCosine,
-        );
-
-        // Clear mappings
-        node_to_index.clear();
-        index_to_node.clear();
-        *next_index = 0;
-
-        // Reset statistics
-        let mut stats = self.stats.write();
-        stats.total_vectors = 0;
-
-        Ok(())
-    }
-
-    /// Normalize a vector to unit length.
-    ///
-    /// Uses the SIMD-dispatched `simd::distance::normalize_f32` kernel
-    /// (AVX-512 → AVX2 → SSE4.2 → NEON → Scalar) — the method keeps
-    /// its `&self` signature for API compatibility but the kernel
-    /// selection is stateless.
-    pub fn normalize_vector(&self, vector: &mut [f32]) {
-        simd::distance::normalize_f32(vector);
-    }
-}
-
-impl Default for KnnIndex {
-    fn default() -> Self {
-        Self::new(128).expect("Failed to create default KNN index")
-    }
-}
-
-/// Property B-tree index for range queries and unique constraints
-///
-/// Maps (label_id, key_id, value) → set of node_ids for fast property-based queries.
-/// Uses BTreeMap for efficient range queries and ordered iteration.
-#[derive(Clone)]
-pub struct PropertyIndex {
-    /// Mapping from (label_id, key_id) to value → set of node_ids
-    property_trees: Arc<RwLock<HashMap<(u32, u32), PropertyIndexTree>>>,
-    /// Statistics
-    stats: Arc<RwLock<PropertyIndexStats>>,
-}
-
-/// Property value for indexing
-#[derive(Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
-pub enum PropertyValue {
-    /// String property value
-    String(String),
-    /// Integer property value
-    Integer(i64),
-    /// Floating point property value
-    Float(f64),
-    /// Boolean property value
-    Boolean(bool),
-    /// Null property value
-    Null,
-}
-
-impl Eq for PropertyValue {}
-
-impl PartialOrd for PropertyValue {
-    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-impl Ord for PropertyValue {
-    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
-        match (self, other) {
-            (PropertyValue::String(a), PropertyValue::String(b)) => a.cmp(b),
-            (PropertyValue::Integer(a), PropertyValue::Integer(b)) => a.cmp(b),
-            (PropertyValue::Float(a), PropertyValue::Float(b)) => {
-                a.partial_cmp(b).unwrap_or(std::cmp::Ordering::Equal)
-            }
-            (PropertyValue::Boolean(a), PropertyValue::Boolean(b)) => a.cmp(b),
-            (PropertyValue::Null, PropertyValue::Null) => std::cmp::Ordering::Equal,
-            // Different types are ordered by variant order
-            (PropertyValue::String(_), _) => std::cmp::Ordering::Less,
-            (PropertyValue::Integer(_), PropertyValue::String(_)) => std::cmp::Ordering::Greater,
-            (PropertyValue::Integer(_), _) => std::cmp::Ordering::Less,
-            (PropertyValue::Float(_), PropertyValue::String(_) | PropertyValue::Integer(_)) => {
-                std::cmp::Ordering::Greater
-            }
-            (PropertyValue::Float(_), _) => std::cmp::Ordering::Less,
-            (PropertyValue::Boolean(_), PropertyValue::Null) => std::cmp::Ordering::Less,
-            (PropertyValue::Boolean(_), _) => std::cmp::Ordering::Greater,
-            (PropertyValue::Null, _) => std::cmp::Ordering::Greater,
-        }
-    }
-}
-
-impl std::hash::Hash for PropertyValue {
-    fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        match self {
-            PropertyValue::String(s) => s.hash(state),
-            PropertyValue::Integer(i) => i.hash(state),
-            PropertyValue::Float(f) => {
-                // Convert f64 to bits for hashing
-                f.to_bits().hash(state);
-            }
-            PropertyValue::Boolean(b) => b.hash(state),
-            PropertyValue::Null => 0.hash(state),
-        }
-    }
-}
-
-/// Statistics for property index
-#[derive(Debug, Clone, Default)]
-pub struct PropertyIndexStats {
-    /// Total number of property entries indexed
-    pub total_entries: u64,
-    /// Number of unique (label_id, key_id) combinations
-    pub indexed_properties: u32,
-    /// Average entries per property
-    pub avg_entries_per_property: f64,
-    /// Memory usage in bytes
-    pub memory_usage_bytes: u64,
-}
-
-impl PropertyIndex {
-    /// Create a new property index
-    pub fn new() -> Self {
-        Self {
-            property_trees: Arc::new(RwLock::new(HashMap::new())),
-            stats: Arc::new(RwLock::new(PropertyIndexStats::default())),
-        }
-    }
-
-    /// Add a property value for a node
-    pub fn add_property(
-        &self,
-        node_id: u64,
-        label_id: u32,
-        key_id: u32,
-        value: PropertyValue,
-    ) -> Result<()> {
-        // Null-key contract (Neo4j-aligned): a null property value means the
-        // property is absent, so it is never indexed. Skipping it keeps the
-        // typed property index free of null-keyed entries — `find_exact(..,
-        // Null)` therefore never matches and legacy null-valued properties
-        // cannot pollute index seeks. See docs/ops/graph-rebuild.md.
-        if value == PropertyValue::Null {
-            return Ok(());
-        }
-
-        let mut trees = self.property_trees.write();
-        let mut stats = self.stats.write();
-
-        let tree = trees.entry((label_id, key_id)).or_default();
-        let bitmap = tree.entry(value).or_default();
-        bitmap.insert(node_id as u32);
-
-        stats.total_entries += 1;
-        stats.indexed_properties = trees.len() as u32;
-        stats.avg_entries_per_property = if stats.indexed_properties > 0 {
-            stats.total_entries as f64 / stats.indexed_properties as f64
-        } else {
-            0.0
-        };
-
-        Ok(())
-    }
-
-    /// Remove a property value for a node
-    pub fn remove_property(
-        &self,
-        node_id: u64,
-        label_id: u32,
-        key_id: u32,
-        value: PropertyValue,
-    ) -> Result<()> {
-        let mut trees = self.property_trees.write();
-        let mut stats = self.stats.write();
-
-        if let Some(tree) = trees.get_mut(&(label_id, key_id)) {
-            if let Some(bitmap) = tree.get_mut(&value) {
-                bitmap.remove(node_id as u32);
-
-                // Remove empty entries
-                if bitmap.is_empty() {
-                    tree.remove(&value);
-                }
-
-                stats.total_entries = stats.total_entries.saturating_sub(1);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// Find nodes with exact property value
-    pub fn find_exact(
-        &self,
-        label_id: u32,
-        key_id: u32,
-        value: PropertyValue,
-    ) -> Result<RoaringBitmap> {
-        let trees = self.property_trees.read();
-
-        if let Some(tree) = trees.get(&(label_id, key_id)) {
-            if let Some(bitmap) = tree.get(&value) {
-                return Ok(bitmap.clone());
-            }
-        }
-
-        Ok(RoaringBitmap::new())
-    }
-
-    /// Find nodes with property value in range
-    pub fn find_range(
-        &self,
-        label_id: u32,
-        key_id: u32,
-        min_value: Option<PropertyValue>,
-        max_value: Option<PropertyValue>,
-    ) -> Result<RoaringBitmap> {
-        let trees = self.property_trees.read();
-        let mut result = RoaringBitmap::new();
-
-        if let Some(tree) = trees.get(&(label_id, key_id)) {
-            let range = match (min_value, max_value) {
-                (Some(min), Some(max)) => tree.range(min..=max),
-                (Some(min), None) => tree.range(min..),
-                (None, Some(max)) => tree.range(..=max),
-                (None, None) => tree.range(..),
-            };
-
-            for (_, bitmap) in range {
-                result |= bitmap;
-            }
-        }
-
-        Ok(result)
-    }
-
-    /// Find nodes with property value greater than threshold
-    pub fn find_greater_than(
-        &self,
-        label_id: u32,
-        key_id: u32,
-        threshold: PropertyValue,
-    ) -> Result<RoaringBitmap> {
-        self.find_range(label_id, key_id, Some(threshold), None)
-    }
-
-    /// Find nodes with property value less than threshold
-    pub fn find_less_than(
-        &self,
-        label_id: u32,
-        key_id: u32,
-        threshold: PropertyValue,
-    ) -> Result<RoaringBitmap> {
-        self.find_range(label_id, key_id, None, Some(threshold))
-    }
-
-    /// Check if a property value exists (for unique constraints)
-    pub fn has_value(&self, label_id: u32, key_id: u32, value: PropertyValue) -> Result<bool> {
-        let trees = self.property_trees.read();
-
-        if let Some(tree) = trees.get(&(label_id, key_id)) {
-            Ok(tree.contains_key(&value))
-        } else {
-            Ok(false)
-        }
-    }
-
-    /// Get all unique values for a property
-    pub fn get_unique_values(&self, label_id: u32, key_id: u32) -> Result<Vec<PropertyValue>> {
-        let trees = self.property_trees.read();
-
-        if let Some(tree) = trees.get(&(label_id, key_id)) {
-            Ok(tree.keys().cloned().collect())
-        } else {
-            Ok(Vec::new())
-        }
-    }
-
-    /// Get statistics
-    pub fn get_stats(&self) -> PropertyIndexStats {
-        self.stats.read().clone()
-    }
-
-    /// Clear all data
-    pub fn clear(&self) -> Result<()> {
-        let mut trees = self.property_trees.write();
-        let mut stats = self.stats.write();
-
-        trees.clear();
-        *stats = PropertyIndexStats::default();
-
-        Ok(())
-    }
-
-    /// Get memory usage estimate
-    pub fn estimate_memory_usage(&self) -> u64 {
-        let trees = self.property_trees.read();
-        let mut total_size = 0;
-
-        for tree in trees.values() {
-            for bitmap in tree.values() {
-                total_size += bitmap.serialized_size() as u64;
-            }
-        }
-
-        total_size
-    }
-
-    /// Check if an index exists for a (label_id, key_id) combination
-    pub fn has_index(&self, label_id: u32, key_id: u32) -> bool {
-        let trees = self.property_trees.read();
-        trees.contains_key(&(label_id, key_id))
-    }
-
-    /// True if at least one property index is registered. Cheap fast-path
-    /// guard (#21): write paths that maintain the typed index can skip the
-    /// per-property × per-label `has_index` loop entirely when no index exists
-    /// — the common case for un-indexed graphs.
-    pub fn has_any_index(&self) -> bool {
-        !self.property_trees.read().is_empty()
-    }
-
-    /// Create an index for a (label_id, key_id) combination
-    /// This initializes an empty index structure. The index will be populated
-    /// as properties are added via add_property().
-    pub fn create_index(&self, label_id: u32, key_id: u32) -> Result<()> {
-        let mut trees = self.property_trees.write();
-        let mut stats = self.stats.write();
-
-        // Create empty index if it doesn't exist
-        if trees.entry((label_id, key_id)).or_default().is_empty() {
-            stats.indexed_properties = trees.len() as u32;
-        }
-
-        Ok(())
-    }
-
-    /// Drop an index for a (label_id, key_id) combination
-    /// This removes all indexed data for this property.
-    pub fn drop_index(&self, label_id: u32, key_id: u32) -> Result<()> {
-        let mut trees = self.property_trees.write();
-        let mut stats = self.stats.write();
-
-        if let Some(tree) = trees.remove(&(label_id, key_id)) {
-            // Update stats: subtract entries from this index
-            let mut removed_entries = 0u64;
-            for bitmap in tree.values() {
-                removed_entries += bitmap.len();
-            }
-            stats.total_entries = stats.total_entries.saturating_sub(removed_entries);
-            stats.indexed_properties = trees.len() as u32;
-            stats.avg_entries_per_property = if stats.indexed_properties > 0 {
-                stats.total_entries as f64 / stats.indexed_properties as f64
-            } else {
-                0.0
-            };
-        }
-
-        Ok(())
-    }
-
-    /// Health check for the property index
-    pub fn health_check(&self) -> Result<()> {
-        let trees = self.property_trees.read();
-        let stats = self.stats.read();
-
-        // Check if the number of indexed properties is reasonable
-        if trees.len() > 1_000_000 {
-            // 1 million max
-            return Err(Error::index("Too many indexed properties"));
-        }
-
-        // Check if the total entries count is reasonable
-        if stats.total_entries > 1_000_000_000 {
-            // 1 billion max
-            return Err(Error::index("Too many entries in property index"));
-        }
-
-        Ok(())
-    }
-}
-
-impl Default for PropertyIndex {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn test_label_index_creation() {
-        let index = LabelIndex::new();
-        let stats = index.get_stats();
-        assert_eq!(stats.total_nodes, 0);
-        assert_eq!(stats.label_count, 0);
-    }
-
-    #[test]
-    fn test_label_index_add_node() {
-        let index = LabelIndex::new();
-
-        index.add_node(1, &[0, 1]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-        index.add_node(3, &[1, 2]).unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_nodes, 3);
-        assert_eq!(stats.label_count, 3);
-
-        let nodes_with_label_0 = index.get_nodes(0).unwrap();
-        assert_eq!(nodes_with_label_0.len(), 2);
-        assert!(nodes_with_label_0.contains(1));
-        assert!(nodes_with_label_0.contains(2));
-    }
-
-    #[test]
-    fn test_label_index_intersection() {
-        let index = LabelIndex::new();
-
-        index.add_node(1, &[0, 1]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-        index.add_node(3, &[1, 2]).unwrap();
-
-        let nodes_with_both = index.get_nodes_with_labels(&[0, 1]).unwrap();
-        assert_eq!(nodes_with_both.len(), 1);
-        assert!(nodes_with_both.contains(1));
-    }
-
-    #[test]
-    fn test_label_index_union() {
-        let index = LabelIndex::new();
-
-        index.add_node(1, &[0, 1]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-        index.add_node(3, &[1, 2]).unwrap();
-
-        let nodes_with_any = index.get_nodes_with_any_labels(&[0, 1]).unwrap();
-        assert_eq!(nodes_with_any.len(), 3);
-    }
-
-    #[test]
-    fn test_label_index_remove_node() {
-        let index = LabelIndex::new();
-
-        index.add_node(1, &[0, 1]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-
-        index.remove_node(1).unwrap();
-
-        let nodes_with_label_0 = index.get_nodes(0).unwrap();
-        assert_eq!(nodes_with_label_0.len(), 1);
-        assert!(nodes_with_label_0.contains(2));
-    }
-
-    #[test]
-    fn test_knn_index_creation() {
-        let index = KnnIndex::new(DEFAULT_VECTORIZER_DIMENSION).unwrap();
-        assert_eq!(index.dimension(), DEFAULT_VECTORIZER_DIMENSION);
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_vectors, 0);
-        assert_eq!(stats.dimension, 128);
-    }
-
-    #[test]
-    fn test_knn_index_add_vector() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let embedding1 = vec![1.0, 0.0, 0.0];
-        let embedding2 = vec![0.0, 1.0, 0.0];
-
-        index.add_vector(1, embedding1).unwrap();
-        index.add_vector(2, embedding2).unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_vectors, 2);
-    }
-
-    #[test]
-    fn test_knn_index_search() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let embedding1 = vec![1.0, 0.0, 0.0];
-        let embedding2 = vec![0.0, 1.0, 0.0];
-        let embedding3 = vec![0.0, 0.0, 1.0];
-
-        index.add_vector(1, embedding1).unwrap();
-        index.add_vector(2, embedding2).unwrap();
-        index.add_vector(3, embedding3).unwrap();
-
-        let query = vec![1.0, 0.0, 0.0];
-        let results = index.search_knn(&query, 2).unwrap();
-
-        // HNSW may return fewer results than k when the index is very small (3 vectors)
-        // This is expected behavior due to the HNSW graph structure
-        assert!(
-            !results.is_empty() && results.len() <= 2,
-            "Should return at least 1 result, at most 2"
-        );
-        assert_eq!(results[0].0, 1); // Most similar to query
-        assert!(results[0].1 > 0.9); // High similarity
-    }
-
-    #[test]
-    fn test_knn_index_search_with_ef_matches_default() {
-        let index = KnnIndex::new(3).unwrap();
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        index.add_vector(2, vec![0.0, 1.0, 0.0]).unwrap();
-        index.add_vector(3, vec![0.0, 0.0, 1.0]).unwrap();
-
-        let query = vec![1.0, 0.0, 0.0];
-        let default_results = index.search_knn(&query, 2).unwrap();
-        let explicit_results = index
-            .search_knn_with_ef(&query, 2, KnnIndex::DEFAULT_EF_SEARCH)
-            .unwrap();
-        assert_eq!(default_results, explicit_results);
-    }
-
-    #[test]
-    fn test_knn_index_search_with_ef_clamps_below_k() {
-        let index = KnnIndex::new(3).unwrap();
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        index.add_vector(2, vec![0.0, 1.0, 0.0]).unwrap();
-        index.add_vector(3, vec![0.0, 0.0, 1.0]).unwrap();
-
-        // ef_search=1 with k=2 must still produce up to k results
-        // because the implementation raises ef to k internally.
-        let results = index
-            .search_knn_with_ef(&vec![1.0, 0.0, 0.0], 2, 1)
-            .unwrap();
-        assert!(!results.is_empty());
-        assert!(results.len() <= 2);
-        assert_eq!(results[0].0, 1);
-    }
-
-    #[test]
-    fn test_knn_index_dimension_mismatch() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let wrong_dimension = vec![1.0, 0.0];
-        let result = index.add_vector(1, wrong_dimension);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_knn_index_remove_vector() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let embedding = vec![1.0, 0.0, 0.0];
-        index.add_vector(1, embedding).unwrap();
-
-        assert!(index.has_vector(1));
-
-        index.remove_vector(1).unwrap();
-
-        assert!(!index.has_vector(1));
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_vectors, 0);
-    }
-
-    #[test]
-    fn test_cosine_similarity() {
-        let _index = KnnIndex::new(3).unwrap();
-
-        let a = vec![1.0, 0.0, 0.0];
-        let b = vec![1.0, 0.0, 0.0];
-        let c = vec![0.0, 1.0, 0.0];
-
-        // Helper function to calculate cosine similarity
-        fn cosine_similarity(a: &[f32], b: &[f32]) -> f32 {
-            let dot_product: f32 = a.iter().zip(b.iter()).map(|(x, y)| x * y).sum();
-            let norm_a: f32 = a.iter().map(|x| x * x).sum::<f32>().sqrt();
-            let norm_b: f32 = b.iter().map(|x| x * x).sum::<f32>().sqrt();
-
-            if norm_a == 0.0 || norm_b == 0.0 {
-                0.0
-            } else {
-                dot_product / (norm_a * norm_b)
-            }
-        }
-
-        // Same vectors should have similarity 1.0
-        let sim_ab = cosine_similarity(&a, &b);
-        assert!((sim_ab - 1.0).abs() < 1e-6);
-
-        // Orthogonal vectors should have similarity 0.0
-        let sim_ac = cosine_similarity(&a, &c);
-        assert!((sim_ac - 0.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_normalize_vector() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let mut vector = vec![3.0, 4.0, 0.0];
-        index.normalize_vector(&mut vector);
-
-        let norm: f32 = vector.iter().map(|&x| x * x).sum::<f32>().sqrt();
-        assert!((norm - 1.0).abs() < 1e-6);
-    }
-
-    #[test]
-    fn test_knn_index_clear() {
-        let mut index = KnnIndex::new(3).unwrap();
-
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        index.add_vector(2, vec![0.0, 1.0, 0.0]).unwrap();
-
-        assert_eq!(index.get_stats().total_vectors, 2);
-
-        index.clear().unwrap();
-
-        assert_eq!(index.get_stats().total_vectors, 0);
-        assert!(index.get_all_nodes().is_empty());
-    }
-
-    #[test]
-    fn test_label_index_clear() {
-        let mut index = LabelIndex::new();
-
-        index.add_node(1, &[0, 1]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-
-        assert_eq!(index.get_stats().total_nodes, 2);
-
-        index.clear().unwrap();
-
-        assert_eq!(index.get_stats().total_nodes, 0);
-        assert_eq!(index.get_stats().label_count, 0);
-    }
 
     #[test]
     fn test_index_manager_creation() {
@@ -1313,485 +175,5 @@ mod tests {
 
         // Health check should pass for empty manager
         manager.health_check().unwrap();
-    }
-
-    #[test]
-    fn test_label_index_estimate_cardinality() {
-        let index = LabelIndex::new();
-
-        // Empty index
-        assert_eq!(index.estimate_cardinality(0), 0);
-
-        // Add some nodes
-        index.add_node(1, &[0]).unwrap();
-        index.add_node(2, &[0]).unwrap();
-        index.add_node(3, &[1]).unwrap();
-
-        assert_eq!(index.estimate_cardinality(0), 2);
-        assert_eq!(index.estimate_cardinality(1), 1);
-        assert_eq!(index.estimate_cardinality(2), 0);
-    }
-
-    #[test]
-    fn test_label_index_has_label() {
-        let index = LabelIndex::new();
-
-        assert!(!index.has_label(0));
-
-        index.add_node(1, &[0]).unwrap();
-        assert!(index.has_label(0));
-        assert!(!index.has_label(1));
-    }
-
-    #[test]
-    fn test_label_index_get_all_labels() {
-        let index = LabelIndex::new();
-
-        assert!(index.get_all_labels().is_empty());
-
-        index.add_node(1, &[0, 1, 2]).unwrap();
-        let labels = index.get_all_labels();
-        assert_eq!(labels.len(), 3);
-        assert!(labels.contains(&0));
-        assert!(labels.contains(&1));
-        assert!(labels.contains(&2));
-    }
-
-    #[test]
-    fn test_label_index_health_check() {
-        let index = LabelIndex::new();
-
-        // Empty index should pass health check
-        index.health_check().unwrap();
-
-        // Add reasonable amount of data
-        for i in 0..1000 {
-            index.add_node(i, &[i as u32 % 10]).unwrap();
-        }
-        index.health_check().unwrap();
-    }
-
-    #[test]
-    fn test_knn_index_new_default() {
-        let index = KnnIndex::new_default(64).unwrap();
-        assert_eq!(index.dimension(), 64);
-    }
-
-    #[test]
-    fn test_knn_index_invalid_dimension() {
-        // Test zero dimension
-        let result = KnnIndex::new(0);
-        assert!(result.is_err());
-
-        // Test too large dimension
-        let result = KnnIndex::new(5000);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_knn_index_search_knn_default() {
-        let index = KnnIndex::new(3).unwrap();
-
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        index.add_vector(2, vec![0.0, 1.0, 0.0]).unwrap();
-
-        let query = vec![1.0, 0.0, 0.0];
-        let results = index.search_knn_default(&query).unwrap();
-        // HNSW may return fewer results than k when the index is very small (2 vectors)
-        // This is expected behavior due to the HNSW graph structure
-        assert!(!results.is_empty() && results.len() <= 2);
-        // Verify the closest result is node 1
-        assert_eq!(results[0].0, 1);
-    }
-
-    #[test]
-    fn test_knn_index_has_vector() {
-        let index = KnnIndex::new(3).unwrap();
-
-        assert!(!index.has_vector(1));
-
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        assert!(index.has_vector(1));
-        assert!(!index.has_vector(2));
-    }
-
-    #[test]
-    fn test_knn_index_get_all_nodes() {
-        let index = KnnIndex::new(3).unwrap();
-
-        assert!(index.get_all_nodes().is_empty());
-
-        index.add_vector(1, vec![1.0, 0.0, 0.0]).unwrap();
-        index.add_vector(2, vec![0.0, 1.0, 0.0]).unwrap();
-
-        let nodes = index.get_all_nodes();
-        assert_eq!(nodes.len(), 2);
-        assert!(nodes.contains(&1));
-        assert!(nodes.contains(&2));
-    }
-
-    #[test]
-    fn test_property_index_creation() {
-        let index = PropertyIndex::new();
-        let stats = index.get_stats();
-        assert_eq!(stats.total_entries, 0);
-        assert_eq!(stats.indexed_properties, 0);
-    }
-
-    #[test]
-    fn test_property_index_add_property() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(1, 0, 1, PropertyValue::Integer(42))
-            .unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_entries, 3);
-        assert_eq!(stats.indexed_properties, 2);
-    }
-
-    #[test]
-    fn test_property_index_remove_property() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_entries, 2);
-
-        index
-            .remove_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_entries, 1);
-    }
-
-    #[test]
-    fn test_property_index_find_exact() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(3, 0, 0, PropertyValue::String("other".to_string()))
-            .unwrap();
-
-        let results = index
-            .find_exact(0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        assert_eq!(results.len(), 2);
-        assert!(results.contains(1));
-        assert!(results.contains(2));
-        assert!(!results.contains(3));
-    }
-
-    #[test]
-    fn test_property_index_find_range() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::Integer(10))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::Integer(20))
-            .unwrap();
-        index
-            .add_property(3, 0, 0, PropertyValue::Integer(30))
-            .unwrap();
-        index
-            .add_property(4, 0, 0, PropertyValue::Integer(40))
-            .unwrap();
-
-        // Range 15-35
-        let results = index
-            .find_range(
-                0,
-                0,
-                Some(PropertyValue::Integer(15)),
-                Some(PropertyValue::Integer(35)),
-            )
-            .unwrap();
-        assert_eq!(results.len(), 2);
-        assert!(results.contains(2));
-        assert!(results.contains(3));
-    }
-
-    #[test]
-    fn test_property_index_find_greater_than() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::Integer(10))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::Integer(20))
-            .unwrap();
-        index
-            .add_property(3, 0, 0, PropertyValue::Integer(30))
-            .unwrap();
-
-        let results = index
-            .find_greater_than(0, 0, PropertyValue::Integer(15))
-            .unwrap();
-        assert_eq!(results.len(), 2);
-        assert!(results.contains(2));
-        assert!(results.contains(3));
-    }
-
-    #[test]
-    fn test_property_index_find_less_than() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::Integer(10))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::Integer(20))
-            .unwrap();
-        index
-            .add_property(3, 0, 0, PropertyValue::Integer(30))
-            .unwrap();
-
-        let results = index
-            .find_less_than(0, 0, PropertyValue::Integer(25))
-            .unwrap();
-        assert_eq!(results.len(), 2);
-        assert!(results.contains(1));
-        assert!(results.contains(2));
-    }
-
-    #[test]
-    fn test_property_index_has_value() {
-        let index = PropertyIndex::new();
-
-        assert!(
-            !index
-                .has_value(0, 0, PropertyValue::String("test".to_string()))
-                .unwrap()
-        );
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        assert!(
-            index
-                .has_value(0, 0, PropertyValue::String("test".to_string()))
-                .unwrap()
-        );
-        assert!(
-            !index
-                .has_value(0, 0, PropertyValue::String("other".to_string()))
-                .unwrap()
-        );
-    }
-
-    #[test]
-    fn test_property_index_get_unique_values() {
-        let index = PropertyIndex::new();
-
-        assert!(index.get_unique_values(0, 0).unwrap().is_empty());
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        index
-            .add_property(2, 0, 0, PropertyValue::String("other".to_string()))
-            .unwrap();
-        index
-            .add_property(3, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-
-        let values = index.get_unique_values(0, 0).unwrap();
-        assert_eq!(values.len(), 2);
-        assert!(values.contains(&PropertyValue::String("test".to_string())));
-        assert!(values.contains(&PropertyValue::String("other".to_string())));
-    }
-
-    #[test]
-    fn test_property_index_clear() {
-        let index = PropertyIndex::new();
-
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        assert_eq!(index.get_stats().total_entries, 1);
-
-        index.clear().unwrap();
-        assert_eq!(index.get_stats().total_entries, 0);
-    }
-
-    #[test]
-    fn test_property_index_estimate_memory_usage() {
-        let index = PropertyIndex::new();
-
-        // Empty index should have minimal memory usage
-        let usage = index.estimate_memory_usage();
-        assert_eq!(usage, 0);
-
-        // Add some data
-        index
-            .add_property(1, 0, 0, PropertyValue::String("test".to_string()))
-            .unwrap();
-        let usage = index.estimate_memory_usage();
-        assert!(usage > 0);
-    }
-
-    #[test]
-    fn test_property_index_health_check() {
-        let index = PropertyIndex::new();
-
-        // Empty index should pass health check
-        index.health_check().unwrap();
-
-        // Add reasonable amount of data
-        for i in 0..1000 {
-            index
-                .add_property(i, i as u32 % 10, 0, PropertyValue::Integer(i as i64))
-                .unwrap();
-        }
-        index.health_check().unwrap();
-    }
-
-    #[test]
-    fn test_property_value_ordering() {
-        // Test ordering of different property value types
-        let values = [
-            PropertyValue::String("a".to_string()),
-            PropertyValue::Integer(1),
-            PropertyValue::Float(1.0),
-            PropertyValue::Boolean(true),
-            PropertyValue::Null,
-        ];
-
-        for i in 0..values.len() {
-            for j in 0..values.len() {
-                if i < j {
-                    assert!(values[i] < values[j]);
-                } else if i > j {
-                    assert!(values[i] > values[j]);
-                } else {
-                    assert_eq!(values[i], values[j]);
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test_property_value_hashing() {
-        use std::collections::HashMap;
-
-        let mut map = HashMap::new();
-
-        map.insert(PropertyValue::String("test".to_string()), 1);
-        map.insert(PropertyValue::Integer(42), 2);
-        map.insert(PropertyValue::Boolean(true), 3);
-
-        assert_eq!(
-            map.get(&PropertyValue::String("test".to_string())),
-            Some(&1)
-        );
-        assert_eq!(map.get(&PropertyValue::Integer(42)), Some(&2));
-        assert_eq!(map.get(&PropertyValue::Boolean(true)), Some(&3));
-        assert_eq!(map.get(&PropertyValue::String("other".to_string())), None);
-    }
-
-    #[test]
-    fn test_label_index_empty_labels() {
-        let index = LabelIndex::new();
-
-        // Test with empty label list
-        let result = index.get_nodes_with_labels(&[]).unwrap();
-        assert!(result.is_empty());
-
-        let result = index.get_nodes_with_any_labels(&[]).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_label_index_nonexistent_labels() {
-        let index = LabelIndex::new();
-
-        // Test with non-existent labels
-        let result = index.get_nodes_with_labels(&[999]).unwrap();
-        assert!(result.is_empty());
-
-        let result = index.get_nodes_with_any_labels(&[999, 998]).unwrap();
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn test_knn_index_search_empty() {
-        let index = KnnIndex::new(3).unwrap();
-
-        let query = vec![1.0, 0.0, 0.0];
-        let results = index.search_knn(&query, 5).unwrap();
-        assert!(results.is_empty());
-    }
-
-    #[test]
-    fn test_knn_index_remove_nonexistent() {
-        let index = KnnIndex::new(3).unwrap();
-
-        // Removing non-existent vector should not error
-        index.remove_vector(999).unwrap();
-
-        let stats = index.get_stats();
-        assert_eq!(stats.total_vectors, 0);
-    }
-
-    #[test]
-    fn add_property_skips_null_value() {
-        let index = PropertyIndex::new();
-        // label_id=1, key_id=1
-        index.create_index(1, 1).unwrap();
-
-        // Adding Null must be a no-op — find_exact returns an empty bitmap.
-        index.add_property(42, 1, 1, PropertyValue::Null).unwrap();
-        let null_hits = index.find_exact(1, 1, PropertyValue::Null).unwrap();
-        assert!(
-            null_hits.is_empty(),
-            "find_exact for Null should return empty bitmap, got {null_hits:?}"
-        );
-
-        // A subsequent non-null add must still be indexed normally.
-        index
-            .add_property(42, 1, 1, PropertyValue::Integer(99))
-            .unwrap();
-        let hits = index.find_exact(1, 1, PropertyValue::Integer(99)).unwrap();
-        assert!(
-            hits.contains(42),
-            "node 42 should appear after non-null add"
-        );
-    }
-
-    #[test]
-    fn has_any_index_reflects_registration() {
-        // #21: the write-path fast-path guard. Empty index => no work.
-        let index = PropertyIndex::new();
-        assert!(!index.has_any_index(), "fresh index has no registrations");
-        index.create_index(1, 1).unwrap();
-        assert!(
-            index.has_any_index(),
-            "after create_index, at least one exists"
-        );
-        index.drop_index(1, 1).unwrap();
-        assert!(!index.has_any_index(), "after drop_index, none remain");
     }
 }
